pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "google-api-python-client>=2.167.0",
    "gspread>=6.2.0",
    "gspread-dataframe>=4.0.0",
    "logfire[fastapi]>=3.14.0",
    "mcp[cli]>=1.6.0",
    "oauth2client>=4.1.3",
    "pydantic-ai>=0.1.2",
    "pyyaml>=6.0.2",
]
